requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
Flask-Bcrypt>=1.0.1
Flask-JWT-Extended>=4.6.0
Flask-Swagger-UI>=4.11.1
APScheduler>=3.10.4
SQLAlchemy>=2.0.27
Werkzeug>=3.0.1
python-dotenv>=1.0.0
pytest>=7.4.3
click>=8.1.7
Jinja2>=3.1.3
itsdangerous>=2.1.2
MarkupSafe>=2.1.5
